requirements.txt
python-dotenv==1.0.1
streamlit==1.39.0
haystack-ai==2.4.1
sentence-transformers==3.1.1
accelerate==1.0.1
torch>=2.3.0
google-generativeai==0.8.5
jinja2==3.1.4